Move fuser to test_jit_fuser
